[fix] 사운드쪽 에러 표시 수정

+++ Assets/@Cosmos/Scripts/Boss/Reaper/ReaperShortDeathAreaPattern.cs
@@ -5,14 +5,14 @@ using System.Collections.Generic;
 public class ReaperShortDeathAreaPattern : IBossAttackPattern
 {
     private GameObject _shortDeathAria;
-    private int _deathAriaTime;
+    private float _deathAriaTime;
     public string PatternName => "ReaperShortDeathAriaPattern";
 
     /// <summary>
     /// 가디언 골렘 패턴1 생성자
     /// </summary>
     /// <param name="poisionAriaPrefab">독 이펙트 프리팹</param>
-    public ReaperShortDeathAreaPattern(GameObject ShortDeathAria, int DeathAriaTime)
+    public ReaperShortDeathAreaPattern(GameObject ShortDeathAria, float DeathAriaTime)
     {
         _shortDeathAria = ShortDeathAria;
         _deathAriaTime = DeathAriaTime;
